перенос проверки на валидность в отдельный модуль
